data_stores_and_pipelines/assignment_1/2025em1100506/ecommerce_seller_recommendation/local/requirements.txt
pyspark==3.5.0
PyYAML==6.0.1
pandas>=2.0.0
matplotlib>=3.7.0
